deleteProdotti
aggiunto il pulsante per rimuovere i prodotti, pienamente funzionante, temporaneamente accessibile a qualsiasi utente

+++ src/main/java/it/uniroma3/siwcatalog/controller/FornitoreController.java
@@ -102,6 +102,9 @@ public class FornitoreController {
     
     @GetMapping(value="/deleteFornitore/{id}")
     public String rimuoviFornitore(@PathVariable("id") Long id, Model model) {
+        if (!(this.fornitoreService.existsById(id))) {
+            return "erroreFornitore.html";
+        }
         this.fornitoreService.deleteFornitore(id);
         model.addAttribute("fornitori",  this.fornitoreService.findAllFornitori());
         return "listaFornitori.html";
